move getOpenSeaExternalUrl to shared package
